requirements.txt
requests>=2.31.0
PyGithub>=1.59.0
python-dateutil>=2.8.2
